Initial conversion to PyQt5 and Python 3.4, moved tests to this repository

+++ src/hinge/crypto/smp.py
@@ -1,4 +1,4 @@
-import CryptoUtils
+from . import CryptoUtils
 import struct
 
 from src.hinge.utils import errors
@@ -204,7 +204,7 @@ class SMP(object):
         return (val >= 2 and val <= self.mod-2)
 
     def hash(self, message):
-        return long(self.crypto.stringHash(message), 16)
+        return int(self.crypto.stringHash(message), 16)
 
 def packList(*items):
     buffer = ''
@@ -241,8 +241,8 @@ def bytesToLong(bytes):
 
 def longToBytes(long):
     bytes = ''
-    while long != 0:
-        bytes = longToByte(long & 0xff) + bytes
+    while int != 0:
+        bytes = longToByte(int & 0xff) + bytes
         long >>= 8
     return bytes
 
@@ -250,7 +250,7 @@ def byteToLong(byte):
     return struct.unpack('B', byte)[0]
 
 def longToByte(long):
-    return struct.pack('B', long)
+    return struct.pack('B', int)
 
 def mulm(x, y, mod):
     return x * y % mod
